requirements.txt
setuptools
tensorflow-gpu==2.1.0
matplotlib
tqdm
